[refactor] DummyData파일 Model 폴더에 추가
